- add LICENSE - DFCache+Extensions -> DFCache+DFExtensions - test targets files cleanup

+++ DFCache Tests/Test Suites/TDFCache+Extensions.m
@@ -10,7 +10,7 @@
  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-#import "DFCache+Extensions.h"
+#import "DFCache+DFExtensions.h"
 #import "DFCache+Tests.h"
 #import "DFTesting.h"
 #import <XCTest/XCTest.h>
@@ -169,7 +169,7 @@
 - (NSString *)_generateCacheName {
     static NSUInteger index = 0;
     index++;
-    return [NSString stringWithFormat:@"_df_test_cache_%i", index];
+    return [NSString stringWithFormat:@"_df_test_cache_%lu", (unsigned long)index];
 }
 
 - (void)_assertContainsObjectsForKeys:(NSArray *)keys objects:(NSDictionary *)objects {
